backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
python-dotenv==1.0.0
crewai==0.11.0
langchain==0.1.0
python-dotenv==1.0.0
openai==1.8.0